chore: code clean up to remove duplication in model/simulation views
