+++ packages/client/hmi-client/src/services/graph.ts
@@ -1,9 +1,54 @@
+/**
+ * Provides graph rendering utilities for computational-like graphs
+ */
 import _ from 'lodash';
+import * as d3 from 'd3';
 import dagre from 'dagre';
 import graphScaffolder, { IGraph, INode, IEdge } from '@graph-scaffolder/index';
 
 export type D3SelectionINode<T> = d3.Selection<d3.BaseType, INode<T>, null, any>;
 export type D3SelectionIEdge<T> = d3.Selection<d3.BaseType, IEdge<T>, null, any>;
+
+const MARKER_VIEWBOX = '-5 -5 10 10';
+const ARROW_PATH = 'M 0,-3.25 L 5 ,0 L 0,3.25';
+
+export const pathFn = d3
+	.line<{ x: number; y: number }>()
+	.x((d) => d.x)
+	.y((d) => d.y);
+
+/**
+ * BasicRenderer with predefined arrow-head markers
+ */
+export abstract class BaseComputionGraph<V, E> extends graphScaffolder.BasicRenderer<V, E> {
+	readonly EDGE_ARROW_ID = 'edge-arrowhead';
+
+	setupDefs() {
+		const svg = d3.select(this.svgEl);
+
+		// Clean up
+		svg.select('defs').selectAll('.edge-marker-end').remove();
+
+		// Arrow defs
+		svg
+			.select('defs')
+			.append('marker')
+			.classed('edge-marker-end', true)
+			.attr('id', this.EDGE_ARROW_ID)
+			.attr('viewBox', MARKER_VIEWBOX)
+			.attr('refX', 2)
+			.attr('refY', 0)
+			.attr('orient', 'auto')
+			.attr('markerWidth', 15)
+			.attr('markerHeight', 15)
+			.attr('markerUnits', 'userSpaceOnUse')
+			.attr('xoverflow', 'visible')
+			.append('svg:path')
+			.attr('d', ARROW_PATH)
+			.style('fill', '#000')
+			.style('stroke', 'none');
+	}
+}
 
 export const runDagreLayout = <V, E>(graphData: IGraph<V, E>): IGraph<V, E> => {
 	const g = new dagre.graphlib.Graph({ compound: true });
