+++ packages/client/hmi-client/src/views/Simulation.vue
@@ -1,9 +1,13 @@
 <script setup lang="ts">
 import { onMounted } from 'vue';
-import graphScaffolder from '@graph-scaffolder/index';
-import { runDagreLayout, D3SelectionINode, D3SelectionIEdge } from '@/services/graph';
+import {
+	runDagreLayout,
+	D3SelectionINode,
+	D3SelectionIEdge,
+	BaseComputionGraph,
+	pathFn
+} from '@/services/graph';
 import { parseSimulationPlan2IGraph } from '@/services/simulation';
-import * as d3 from 'd3';
 
 interface NodeData {
 	boxType: string;
@@ -11,40 +15,7 @@ interface NodeData {
 }
 interface EdgeData {}
 
-const MARKER_VIEWBOX = '-5 -5 10 10';
-const ARROW = 'M 0,-3.25 L 5 ,0 L 0,3.25';
-const pathFn = d3
-	.line<{ x: number; y: number }>()
-	.x((d) => d.x)
-	.y((d) => d.y);
-
-class SimulationPlanRenderer extends graphScaffolder.BasicRenderer<NodeData, EdgeData> {
-	setupDefs() {
-		const svg = d3.select(this.svgEl);
-
-		// Clean up
-		svg.select('defs').selectAll('.edge-marker-end').remove();
-
-		// Arrow defs
-		svg
-			.select('defs')
-			.append('marker')
-			.classed('edge-marker-end', true)
-			.attr('id', 'arrowhead')
-			.attr('viewBox', MARKER_VIEWBOX)
-			.attr('refX', 2)
-			.attr('refY', 0)
-			.attr('orient', 'auto')
-			.attr('markerWidth', 15)
-			.attr('markerHeight', 15)
-			.attr('markerUnits', 'userSpaceOnUse')
-			.attr('xoverflow', 'visible')
-			.append('svg:path')
-			.attr('d', ARROW)
-			.style('fill', '#000')
-			.style('stroke', 'none');
-	}
-
+class SimulationPlanRenderer extends BaseComputionGraph<NodeData, EdgeData> {
 	renderNodes(selection: D3SelectionINode<NodeData>) {
 		selection
 			.append('rect')
@@ -76,7 +47,7 @@ class SimulationPlanRenderer extends graphScaffolder.BasicRenderer<NodeData, Edg
 			.style('fill', 'none')
 			.style('stroke', '#000')
 			.style('stroke-width', 2)
-			.attr('marker-end', 'url(#arrowhead)');
+			.attr('marker-end', `url(#${this.EDGE_ARROW_ID})`);
 	}
 }
